+++ src/main/java/com/gmsbank/repository/TiposContaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TiposContaRepository extends JpaRepository<TiposConta, Integer> {
+public interface TiposContaRepository extends JpaRepository<TiposConta, Long> {
 }
